requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
apscheduler==3.10.4
    # via -r requirements.in
async-timeout==4.0.3
    # via redis
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
fastapi==0.115.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==3.0.0
    # via redis
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
numpy==2.0.2
    # via -r requirements.in
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.23.4
    # via pydantic
pytest==8.3.3
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==0.24.0
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pytz==2024.2
    # via apscheduler
pyyaml==6.0.2
    # via uvicorn
redis[hiredis]==5.0.8
    # via -r requirements.in
six==1.16.0
    # via apscheduler
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.35
    # via
    #   -r requirements.in
    #   alembic
starlette==0.38.5
    # via fastapi
tomli==2.0.1
    # via pytest
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   uvicorn
tzlocal==5.2
    # via apscheduler
uvicorn[standard]==0.30.6
    # via -r requirements.in
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
